Cleaned up the infinite reactor simulation function in preparation for temperature based time steps. Also changed Real to be a long double as delayed precursors seemed to be getting truncated.

+++ source/EnumsAndFunctions.h
@@ -16,8 +16,8 @@
 #ifndef MATERIALENUM_H
 #define MATERIALENUM_H
 
-typedef double Real;
-typedef double Dimension;
+typedef long double Real;
+typedef long double Dimension;
 
 enum Materials
 {
